+++ main/src/ca/uwaterloo/flix/language/phase/Monomorpher.scala
@@ -563,11 +563,8 @@ object Monomorpher {
       }.get // This is safe since the last case can always match.
 
     case LoweredAst.Expr.JvmReflection(exp, tpe, eff, loc) =>
-      // Turn `$JVM_TYPE_OF_PROXY$(e: Proxy[t])` into `e; JvmType.Xyz`
-      // where `Xyz` is a reflection of `t`.
       val e = specializeExp(exp, env0, subst)
-      val jvmType = reflectProxyType(e.tpe, loc)
-      MonoAst.Expr.Stm(e, jvmType, tpe, eff, loc)
+      MonoAst.Expr.JvmReflection(e, subst(tpe), subst(eff), loc)
 
     case LoweredAst.Expr.VectorLit(exps, tpe, eff, loc) =>
       val es = exps.map(specializeExp(_, env0, subst))
@@ -588,7 +585,7 @@ object Monomorpher {
     case LoweredAst.Expr.Cast(exp, _, _, tpe, eff, loc) =>
       // Drop the declaredType and declaredEff.
       val e = specializeExp(exp, env0, subst)
-      mkCast(e, subst(tpe), subst(eff), loc)
+      MonoAst.Expr.Cast(e, subst(tpe), subst(eff), loc)
 
     case LoweredAst.Expr.TryCatch(exp, rules, tpe, eff, loc) =>
       val e = specializeExp(exp, env0, subst)
@@ -620,112 +617,6 @@ object Monomorpher {
       val methods = methods0.map(specializeJvmMethod(_, env0, subst))
       MonoAst.Expr.NewObject(name, clazz, subst(tpe), subst(eff), methods, loc)
 
-  }
-
-  /**
-    * Returns the `JvmType` representing `tpe`.
-    *
-    * E.g. `String` returns `JvmType.JvmObject: JvmType \ {}`.
-    *
-    * N.B.: `tpe` must be `Proxy[t]` for some `t`.
-    */
-  private def reflectProxyType(tpe: Type, loc: SourceLocation): MonoAst.Expr = {
-    val innerType = tpe match {
-      case Type.Apply(Type.Cst(_, _), arg, _) => arg
-      case _ => throw InternalCompilerException(s"Unexpected type '$tpe'", tpe.loc)
-    }
-    val constructorName = innerType match {
-      case Type.Char => "JvmChar"
-      case Type.Bool => "JvmBool"
-      case Type.Int8 => "JvmInt8"
-      case Type.Int16 => "JvmInt16"
-      case Type.Int32 => "JvmInt32"
-      case Type.Int64 => "JvmInt64"
-      case Type.Float32 => "JvmFloat32"
-      case Type.Float64 => "JvmFloat64"
-      case Type.Cst(_, _) => "JvmObject"
-      case Type.Apply(_, _, _) => "JvmObject"
-      case Type.Var(_, _) => throw InternalCompilerException(s"Unexpected type '$innerType'", innerType.loc)
-      case Type.Alias(_, _, _, _) => throw InternalCompilerException(s"Unexpected type '$innerType'", innerType.loc)
-      case Type.AssocType(_, _, _, _) => throw InternalCompilerException(s"Unexpected type '$innerType'", innerType.loc)
-      case Type.JvmToType(_, _) => throw InternalCompilerException(s"Unexpected type '$innerType'", innerType.loc)
-      case Type.JvmToEff(_, _) => throw InternalCompilerException(s"Unexpected type '$innerType'", innerType.loc)
-      case Type.UnresolvedJvmType(_, _) => throw InternalCompilerException(s"Unexpected type '$innerType'", innerType.loc)
-    }
-    val tag = Symbol.mkCaseSym(Symbol.JvmType, Name.Ident(constructorName, loc))
-    MonoAst.Expr.ApplyAtomic(AtomicOp.Tag(tag), Nil, Type.mkEnum(Symbol.JvmType, Kind.Star, loc), Type.Pure, loc)
-  }
-
-  /**
-    * Returns the cast of `e` to `tpe` and `eff`.
-    *
-    * If `exp` and `tpe` is bytecode incompatible, a runtime crash is inserted to appease the
-    * bytecode verifier.
-    */
-  private def mkCast(exp: MonoAst.Expr, tpe: Type, eff: Type, loc: SourceLocation): MonoAst.Expr = {
-    (exp.tpe, tpe) match {
-      case (Type.Char, Type.Char) => MonoAst.Expr.Cast(exp, tpe, eff, loc)
-      case (Type.Char, Type.Int32) => MonoAst.Expr.Cast(exp, tpe, eff, loc)
-      case (Type.Bool, Type.Bool) => MonoAst.Expr.Cast(exp, tpe, eff, loc)
-      case (Type.Int8, Type.Int8) => MonoAst.Expr.Cast(exp, tpe, eff, loc)
-      case (Type.Int16, Type.Int16) => MonoAst.Expr.Cast(exp, tpe, eff, loc)
-      case (Type.Int32, Type.Int32) => MonoAst.Expr.Cast(exp, tpe, eff, loc)
-      case (Type.Int64, Type.Int64) => MonoAst.Expr.Cast(exp, tpe, eff, loc)
-      case (Type.Float32, Type.Float32) => MonoAst.Expr.Cast(exp, tpe, eff, loc)
-      case (Type.Float64, Type.Float64) => MonoAst.Expr.Cast(exp, tpe, eff, loc)
-      case (x, y) if !isPrimType(x) && !isPrimType(y) => MonoAst.Expr.Cast(exp, tpe, eff, loc)
-      case (x, y) =>
-        val crash = MonoAst.Expr.ApplyAtomic(AtomicOp.CastError(erasedString(x), erasedString(y)), Nil, tpe, eff, loc)
-        MonoAst.Expr.Stm(exp, crash, tpe, eff, loc)
-    }
-  }
-
-  /**
-    * Returns `true` if `tpe` is a primitive type.
-    *
-    * N.B.: `tpe` must be normalized.
-    */
-  private def isPrimType(tpe: Type): Boolean = tpe match {
-    case Type.Char => true
-    case Type.Bool => true
-    case Type.Int8 => true
-    case Type.Int16 => true
-    case Type.Int32 => true
-    case Type.Int64 => true
-    case Type.Float32 => true
-    case Type.Float64 => true
-    case Type.Cst(_, _) => false
-    case Type.Apply(_, _, _) => false
-    case Type.Var(_, _) => throw InternalCompilerException(s"Unexpected type '$tpe'", tpe.loc)
-    case Type.Alias(_, _, _, _) => throw InternalCompilerException(s"Unexpected type '$tpe'", tpe.loc)
-    case Type.AssocType(_, _, _, _) => throw InternalCompilerException(s"Unexpected type '$tpe'", tpe.loc)
-    case Type.JvmToType(_, _) => throw InternalCompilerException(s"Unexpected type '$tpe'", tpe.loc)
-    case Type.JvmToEff(_, _) => throw InternalCompilerException(s"Unexpected type '$tpe'", tpe.loc)
-    case Type.UnresolvedJvmType(_, _) => throw InternalCompilerException(s"Unexpected type '$tpe'", tpe.loc)
-  }
-
-  /**
-    * Returns the erased string representation of `tpe`
-    *
-    * N.B.: `tpe` must be normalized.
-    */
-  private def erasedString(tpe: Type): String = tpe match {
-    case Type.Char => "Char"
-    case Type.Bool => "Bool"
-    case Type.Int8 => "Int8"
-    case Type.Int16 => "Int16"
-    case Type.Int32 => "Int32"
-    case Type.Int64 => "Int64"
-    case Type.Float32 => "Float32"
-    case Type.Float64 => "Float64"
-    case Type.Cst(_, _) => "Object"
-    case Type.Apply(_, _, _) => "Object"
-    case Type.Var(_, _) => throw InternalCompilerException(s"Unexpected type '$tpe'", tpe.loc)
-    case Type.Alias(_, _, _, _) => throw InternalCompilerException(s"Unexpected type '$tpe'", tpe.loc)
-    case Type.AssocType(_, _, _, _) => throw InternalCompilerException(s"Unexpected type '$tpe'", tpe.loc)
-    case Type.JvmToType(_, _) => throw InternalCompilerException(s"Unexpected type '$tpe'", tpe.loc)
-    case Type.JvmToEff(_, _) => throw InternalCompilerException(s"Unexpected type '$tpe'", tpe.loc)
-    case Type.UnresolvedJvmType(_, _) => throw InternalCompilerException(s"Unexpected type '$tpe'", tpe.loc)
   }
 
   /**
